Add start date field to slate model for calculating current week
Add a new field to the slate model, which should be set to the start
date of broadcasting. This date is then used as a reference to
calculate the "broadcast week" or "week from start", used by the
frontend to figure out if the current week is odd or even.