Disable image dimension tracking for shows
As far as I can ascertain we never use this, and it's causing the
primary show query to take about 7 seconds on average while it
fetches and reads multiple images in order to calculate the dimensions.
There may be a method of forcing the model to re-save whenever the
dimensions are recalculated but I don't know what it is.
Fields themselves stay for now but they may be removed once it's
confirmed that they are unused.